pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "tundri"
version = "1.3.0"
description = "Drop, create and alter Snowflake objects and set permissions with Permifrost"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Gemma Analytics", email = "bijan.soltani@gemmaanalytics.com"}
]
keywords = ["snowflake", "permifrost", "database", "ddl"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "permifrost @ git+https://github.com/Gemma-Analytics/permifrost@main",
    "pyyaml",
    "snowflake-connector-python",
    "python-dotenv",
    "rich",
]

[project.scripts]
tundri = "tundri.cli:main"

[project.urls]
Homepage = "https://github.com/Gemma-Analytics/tundri"
Repository = "https://github.com/Gemma-Analytics/tundri"
Issues = "https://github.com/Gemma-Analytics/tundri/issues"

[dependency-groups]
dev = [
    "black>=24.8.0",
    "pytest>=8.3.5",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
